Add Mac ARM support

+++ src/main/java/jace/core/TimedDevice.java
@@ -30,8 +30,8 @@ import jace.config.ConfigurableField;
 public abstract class TimedDevice extends Device {
     // From the holy word of Sather 3:5 (Table 3.1) :-)
     // This average speed averages in the "long" cycles
-    public static final long NTSC_1MHZ = 1020484L;
-    public static final long PAL_1MHZ = 1015625L;
+    public static final int NTSC_1MHZ = 1020484;
+    public static final int PAL_1MHZ = 1015625;
     public static final long SYNC_FREQ_HZ = 60;
     public static final double NANOS_PER_SECOND = 1000000000.0;
     public static final long NANOS_PER_MILLISECOND = 1000000L;
